Moved transformer operation to beginning (#18)
no-issue

This is to ensure that the filter is transformed before any other merging/expansions occur. The idea behind this is that the queryString
can be end-user defined, wheras the expansions, overrides, defaults are
defined by the developer. The transformer option is to enable the developer
to clean up the queryString passed by the user