Changed regex operators to only work with STRINGs
refs: https://github.com/TryGhost/NQL/issues/22

- The main use case for regex operators is doing type-ahead type matches on values.
These kinds of values will start with a single char, contain spaces, and contain characters
that should be interpretted literally.

- To support this, the right value type in NQL is STRING, not LITERAL.
LITERALs are designed for matching tag slugs and similar computed values.
This commit changes the behaviour completely to only support STRING.

- We also need to be super careful about how we handle regex characters and strings with SQL
There I've added some more tests to demonstrate that the right thing is happening.
I reordered some regex chars as the order ${} means something in JS template literals so this
made it easier to test.
And I added a few test cases for some scary-looking sqli attempts which should not work thanks
to our usage of knex.

NOTE: In testing this, I realised we don't need to be using whereRaw, as knex does actually have
whereLike and whereILike which work correctly.