Added support for single character strings (#91)
ref https://github.com/TryGhost/NQL/issues/20

I -believe- this works all right. It's at the end of the rules so it should allow for matching on just about anything else practical, then allow us a chance at matching a single character string, before throwing a Lexical Error.

I've added negation for all special chars except for + and - as you ought to be able to do 'post:a+other' and 'post:a-something'.